Fix to resolve problem with SD card.
Here is description
https://3dprinting.stackexchange.com/questions/962/printer-randomly-moves-to-home-during-printing-then-resumes-as-normal
And I see the same on my TEVO MichelAngelo. When I checked on SD CRC
check, observed errors greatly reduced. The other options are for the
case for additional sure. Even with SD CRC check nozzles runing away to the
zero become very rare, but still exists.